Extract internal/rowreaderiterator.h and .cc
